Do not change file extension when resizing images.
odm_texture uses the list.txt file created for bundler.
The file contains the list of resized images.
odm_texture uses this filenames to get the original images.
If the name of the original images and the name of the resized
images is not the same, odm_texture does not find the original images.