Altered to work with Pictures without these tags.
From the tutorial of pyexif2:
'''
As a handy shortcut, you can always assign a value for a given key regardless of whether it’s already present in the metadata. If a tag was present, its value is overwritten. If the tag was not present, one is created and its value is set:

>>> metadata[key] = value

'''
http://tilloy.net/dev/pyexiv2/tutorial.html