Linked contour map generation with ODM DEM model
This update links the grass contour code to the ODM project DEM model generated with PDAL, which is located in the project folder /odm_georeferencing/odm_georeferencing_model_dem.tif. 
It also enables contour and relief maps to be created across regions as it ignores the Location PROJ_INFO with the -o flag.