Fix Buildtime Race Condition
The `mv` in `SuperBuild` for `lasmerge` will copy `lasmerge` to a file
called `SuperBuild/install/bin` **not**
`SuperBuild/install/bin/lasmerge` when the `bin` folder has not been
created. This is the case when building with 2 jobs (`-j2`) so builds
fail when moving onto OpenCV which requires that `bin` be a directory.

* Change `mv` to more reliable `install` command for `lasmerge`.

Signed-off-by: Daniel Llewellyn <daniel@snapcraft.ninja>